cmake list update and prepush test update
CMake list update: Separated out the making a lib directory and copying over the libgmml file. This is done to prevent the copy from failing when the lib dir already exists. Am unsure if that is exactly what was happening but I wanted to make sure it would not occur

prepush update: explicitly state that we are using GEMS make.sh to run the test, also we are now wrapping when making because we must ensure that gmml is wrapped since the prepush tests on gmml level end up calling GEMS tests.