clang tidy updates
also include script that we can run that performs the needed workarounds. We have to run the clang tidy a couple times to make sure all checks are done. Eventually I will be saving the output to a file so we can diff to it to programatically check when the edits are complete