Bugfix for multiple branching from sialic acids
When more than one residue is attached to the glycerol sidechain of sialic acid, e.g. 2-8 and 2-7 linkages to the same Neu5Ac, then more meta data is found as a 2-7 linkage should also be able to rotate the dihedrals for the e.g. C9-C8 bond, but if there is a substitient there then that linkage controls those dihedrals. This exposed a bug in that I was allowing <= vector size instead of < vector size when inserting. Note this whole design is very poor, I should not be inserting using indices like this.