more changes to prep for tooling time my dude and small fix
allow cmake to skip whole blocks that end up allowing makefile to wrap code, needs to be done in order to prevent tooling from running on the generated cxx file that is used to wrap in python. If we dont do this is increases the time the tooling takes by thousands of percent...... not goode

also fixed problem where the generated compile commands wouldnt id our files that we consider "system files" as such. This is important cause makes tooling easier