Cannot create a cds Class inside, need to pass in pre-constructed class.
If you call one of these create methods from e.g. PdbMolecule, the residue you construct is  cds::Residue, not PdbResidue. But you can construct the PdbResidue unique_ptr in the function call and move it into the cds::Residue vector, and it will be (and be downcast-able to) the PdbResidue type.