Merging the Abstract Residue into cds::Residue. Caused Issues.
Need to first convert the InternalPrograms carb builder to use the new carb builder class. As well as the old tests and probably the Internal programs gp builder. Then move the condensedSequence class into cds and delete the dependence of old Assembly on the new classes. i.e. delete AssemblyBuilder