Bugfix for carb builder segfaulting in some cases.
Instantiating the carb builder but not setting good default torsion angles or resolving overlaps will cause a bug later if GenerateSingle3DStructureDefaultFiles() isn't called before GenerateSpecific3DStructure. I did this to be efficient when just validating a sequence, but it's clearly a bad idea. This broke gems as GenerateSpecific3DStructure() will sometimes segfault, and so our "database" might be full of bad structures. If a users builds a glycan where the default has already been generated, you will get garbage.