bugfix: this is hacky because our logic for labeling breaks on ano-ano
More detail of the underlying issue here; https://github.com/GLYCAM-Web/gmml/issues/165. For now I've added this code to add the anomericity of the rightmost ano-ano sugar back into the "orderedSequence" generated by gmml. Gems uses the orderedSequence and not the user supplied one, which is why it was not able to build ano-ano, while gmml could.