Bugfix: Normalizing bond lengths then create missing tetrahedral atom
The created oxygens in the 1** residues were a bit off as the C-h bond length is shorter, resulting in the O being a bit wonky. I normalize the lengths of the C-O and C-C and C-H bonds now so the new atom gets put in a better place. This affects a lot of things though because dihedral uses this class if the H atom is missing for a psi dihedral...