Feature: Switched to Lib! The tests all pass.
parameter manager now loads the lib file we created in test 027. Order of atoms means all the output files changed in the tests. With extra lib entries we are able to build more sugars in the tests too, and discovered missing metadata for dihedrals. That metadata was also messed up by clang formatter, so I've added a line for clang to leave it alone.